pyproject.toml
[project]
name = "tds-project-1-virtual-ta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-restful>=0.3.10",
    "google-genai>=1.20.0",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
